refactor: hooks/query 폴더 생성 및 분리
