+++ frontend/src/hooks/query/useGuestSpaces.ts
