+++ frontend/src/hooks/query/useManagerSpaces.ts
