+++ frontend/src/pages/GuestReservation/GuestReservation.tsx
@@ -9,7 +9,7 @@ import PageHeader from 'components/PageHeader/PageHeader';
 import ReservationListItem from 'components/ReservationListItem/ReservationListItem';
 import MESSAGE from 'constants/message';
 import { HREF } from 'constants/path';
-import useGuestReservations from 'hooks/useGuestReservations';
+import useGuestReservations from 'hooks/query/useGuestReservations';
 import useInput from 'hooks/useInput';
 import { GuestMapState } from 'pages/GuestMap/GuestMap';
 import { MapItem, Reservation, ScrollPosition, Space } from 'types/common';
