+++ frontend/src/pages/ManagerMain/ManagerMain.tsx
@@ -20,8 +20,8 @@ import Panel from 'components/Panel/Panel';
 import ReservationListItem from 'components/ReservationListItem/ReservationListItem';
 import MESSAGE from 'constants/message';
 import PATH, { HREF } from 'constants/path';
-import useManagerMaps from 'hooks/useManagerMaps';
-import useManagerReservations from 'hooks/useManagerReservations';
+import useManagerMaps from 'hooks/query/useManagerMaps';
+import useManagerReservations from 'hooks/query/useManagerReservations';
 import { Order, Reservation } from 'types/common';
 import { ErrorResponse, MapItemResponse } from 'types/response';
 import { formatDate } from 'utils/datetime';
