+++ frontend/src/pages/ManagerMapEditor/ManagerMapEditor.tsx
@@ -8,10 +8,10 @@ import Header from 'components/Header/Header';
 import Layout from 'components/Layout/Layout';
 import MESSAGE from 'constants/message';
 import PATH, { HREF } from 'constants/path';
+import useManagerMap from 'hooks/query/useManagerMap';
+import useManagerSpaces from 'hooks/query/useManagerSpaces';
 import useInputs from 'hooks/useInputs';
 import useListenManagerMainState from 'hooks/useListenManagerMainState';
-import useManagerMap from 'hooks/useManagerMap';
-import useManagerSpaces from 'hooks/useManagerSpaces';
 import { ManagerSpace, MapDrawing, MapElement, SpaceArea } from 'types/common';
 import { ErrorResponse } from 'types/response';
 import { createMapImageSvg } from 'utils/map';
