+++ frontend/src/pages/ManagerSpaceEditor/ManagerSpaceEditor.tsx
@@ -14,9 +14,9 @@ import Header from 'components/Header/Header';
 import Layout from 'components/Layout/Layout';
 import { BOARD } from 'constants/editor';
 import MESSAGE from 'constants/message';
+import useManagerMap from 'hooks/query/useManagerMap';
+import useManagerSpaces from 'hooks/query/useManagerSpaces';
 import useListenManagerMainState from 'hooks/useListenManagerMainState';
-import useManagerMap from 'hooks/useManagerMap';
-import useManagerSpaces from 'hooks/useManagerSpaces';
 import { ManagerSpace, MapDrawing, SpaceArea } from 'types/common';
 import { SpaceEditorMode as Mode } from 'types/editor';
 import { ErrorResponse } from 'types/response';
