+++ frontend/src/pages/ManagerSpaceEditor/units/Preset.tsx
@@ -7,8 +7,8 @@ import Button from 'components/Button/Button';
 import IconButton from 'components/IconButton/IconButton';
 import Select from 'components/Select/Select';
 import MESSAGE from 'constants/message';
+import usePresets from 'hooks/query/usePreset';
 import useInput from 'hooks/useInput';
-import usePresets from 'hooks/usePreset';
 import { Preset as PresetType } from 'types/common';
 import { ErrorResponse } from 'types/response';
 import { SpaceFormValue } from '../data';
